AddBookPage refactoring: AuthorSelection, BookDetails, BookMetadata components
